Don't use randomSentence in ask if the given key was unknown.
Due to a flaw in the random key derivation function, keys were sometimes
generated that couldn't be used to derive a random sentence. To make matters
even worse, the Core functions don't return a new StdGen the way they should.

This would result in an endless loop when a bad key coincided with certain
StdGens, as no new StdGen would ever be derived, and ask would eternally
recurse to randomSentence and back again.

For a long-term solution, randomKey should be fixed to never generate invalid
keys and all Core functions that do something random should be changed to
return a new StdGen.

This fix prevents the infinite loop from occurring until matters are fixed.
One could also argue that if asked to generate a sentence based on X,
returning an empty string and let the user take care from there is better than
to generate something completely random and not tell the user about what
happened.